#048 - Added authorization to all (protected) routes

+++ api/api-server.py
@@ -123,79 +123,52 @@ def register():
 # Return all my fav sites
 @app.route('/favorites', methods=['GET'])
 def get_favorites():
-    auth_header = request.headers.get('Authorization')
-    if auth_header.startswith('Bearer '):
-        try:
-            encoded_jwt = auth_header[7:]
-            decoded_jwt = jwt.decode(encoded_jwt, os.getenv('JWT_KEY'), algorithms=['HS256'])
-            if 'username' not in decoded_jwt:
-                return jsonify({"error":"Unauthorized access"}), 403
-            user_id = decoded_jwt['username']
-            favorites = favoriteSites.aggregate([
-                { "$addFields": { "obj_site_id": { "$toObjectId": "$site_id" }}},
-                { "$lookup": { "from": "sites", "localField": "obj_site_id", "foreignField": "_id", "as": "sitesLookup" }},
-                { "$unwind": "$sitesLookup" },
-                { "$match": { "user_id": str(user_id) }},
-                { "$project": { "tag": 1, "views": 1, "lastViewedOn": 1, "dateAdded": 1, "url": "$sitesLookup.url" }},
-                { "$sort": { "tag": 1 }}
-            ])
-            return jsonify(favorites)
-        except Exception as e:
-            print("JWT decode error: ", e)
-            return jsonify({"error":"Unauthorized access"}), 403
-    return jsonify({"error":"Unauthorized access"}), 403
+    user_id = check_authorization(request)
+    if user_id is None:
+        return jsonify({"message":"Unauthorized access"}), 403
+    favorites = favoriteSites.aggregate([
+        { "$addFields": { "obj_site_id": { "$toObjectId": "$site_id" }}},
+        { "$lookup": { "from": "sites", "localField": "obj_site_id", "foreignField": "_id", "as": "sitesLookup" }},
+        { "$unwind": "$sitesLookup" },
+        { "$match": { "user_id": str(user_id) }},
+        { "$project": { "tag": 1, "views": 1, "lastViewedOn": 1, "dateAdded": 1, "url": "$sitesLookup.url" }},
+        { "$sort": { "tag": 1 }}
+    ])
+    return jsonify(favorites)
 
 # Return most recent
 @app.route('/recent', methods=['GET'])
 def get_recent():
-    auth_header = request.headers.get('Authorization')
-    if auth_header.startswith('Bearer '):
-        try:
-            encoded_jwt = auth_header[7:]
-            decoded_jwt = jwt.decode(encoded_jwt, os.getenv('JWT_KEY'), algorithms=['HS256'])
-            if 'username' not in decoded_jwt:
-                return jsonify({"error":"Unauthorized access"}), 403
-            user_id = decoded_jwt['username']
-            recent = favoriteSites.aggregate([
-                { "$addFields": { "obj_site_id": { "$toObjectId": "$site_id" }}},
-                { "$lookup": { "from": "sites", "localField": "obj_site_id", "foreignField": "_id", "as": "sitesLookup" }},
-                { "$unwind": "$sitesLookup" },
-                { "$match": { "user_id": str(user_id) }},
-                { "$project": { "tag": 1, "views": 1, "lastViewedOn": 1, "dateAdded": 1, "url": "$sitesLookup.url" }},
-                { "$sort": { "lastViewedOn": -1 }},
-                { "$limit": 10 }
-            ])
-            return jsonify(recent)
-        except Exception as e:
-            print("JWT decode error: ", e)
-            return jsonify({"error":"Unauthorized access"}), 403
-    return jsonify({"error":"Unauthorized access"}), 403
+    user_id = check_authorization(request)
+    if user_id is None:
+        return jsonify({"message":"Unauthorized access"}), 403
+    recent = favoriteSites.aggregate([
+        { "$addFields": { "obj_site_id": { "$toObjectId": "$site_id" }}},
+        { "$lookup": { "from": "sites", "localField": "obj_site_id", "foreignField": "_id", "as": "sitesLookup" }},
+        { "$unwind": "$sitesLookup" },
+        { "$match": { "user_id": str(user_id) }},
+        { "$project": { "tag": 1, "views": 1, "lastViewedOn": 1, "dateAdded": 1, "url": "$sitesLookup.url" }},
+        { "$sort": { "lastViewedOn": -1 }},
+        { "$limit": 10 }
+    ])
+    return jsonify(recent)
 
 # Return most viewed
 @app.route('/most-viewed', methods=['GET'])
 def get_most_viewed():
-    auth_header = request.headers.get('Authorization')
-    if auth_header.startswith('Bearer '):
-        try:
-            encoded_jwt = auth_header[7:]
-            decoded_jwt = jwt.decode(encoded_jwt, os.getenv('JWT_KEY'), algorithms=['HS256'])
-            if 'username' not in decoded_jwt:
-                return jsonify({"error":"Unauthorized access"}), 403
-            user_id = decoded_jwt['username']
-            mostViewed = favoriteSites.aggregate([
-                { "$addFields": { "obj_site_id": { "$toObjectId": "$site_id" }}},
-                { "$lookup": { "from": "sites", "localField": "obj_site_id", "foreignField": "_id", "as": "sitesLookup" }},
-                { "$unwind": "$sitesLookup" },
-                { "$match": { "user_id": str(user_id) }},
-                { "$project": { "tag": 1, "views": 1, "lastViewedOn": 1, "dateAdded": 1, "url": "$sitesLookup.url" }},
-                { "$sort": { "views": -1 }},
-                { "$limit": 10 }
-            ])
-            return jsonify(mostViewed)
-        except Exception as e:
-            print("JWT decode error: ", e)
-            return jsonify({"error":"Unauthorized access"}), 403
-    return jsonify({"error":"Unauthorized access"}), 403
+    user_id = check_authorization(request)
+    if user_id is None:
+        return jsonify({"message":"Unauthorized access"}), 403
+    mostViewed = favoriteSites.aggregate([
+        { "$addFields": { "obj_site_id": { "$toObjectId": "$site_id" }}},
+        { "$lookup": { "from": "sites", "localField": "obj_site_id", "foreignField": "_id", "as": "sitesLookup" }},
+        { "$unwind": "$sitesLookup" },
+        { "$match": { "user_id": str(user_id) }},
+        { "$project": { "tag": 1, "views": 1, "lastViewedOn": 1, "dateAdded": 1, "url": "$sitesLookup.url" }},
+        { "$sort": { "views": -1 }},
+        { "$limit": 10 }
+    ])
+    return jsonify(mostViewed)
 
 # View site
 @app.route('/view-site', methods=['POST'])
@@ -227,7 +200,9 @@ def get_site_info():
 # Add new favorite site
 @app.route('/add-favorite', methods=['POST'])
 def add_favorite():
-    # Authorization
+    user_id = check_authorization(request)
+    if user_id is None:
+        return jsonify({"message":"Unauthorized access"}), 403
     try:
         mySite = sites.find_one({"url": request.json['url']})
         if mySite is None:
@@ -237,12 +212,12 @@ def add_favorite():
             })
         site_id = str(sites.find_one({"url": request.json['url']})['_id'])
         # Duplicate check not working ???
-        duplicates_count = lists.count_documents({"user_id": request.json['userID'], "site_id": site_id, "tag": request.json['tag']})
+        duplicates_count = lists.count_documents({"user_id": user_id, "site_id": site_id, "tag": request.json['tag']})
         if duplicates_count > 0:
             return jsonify({"message":"You have already saved this site with this tag"}) # error code
         else:
             favoriteSites.insert_one({
-                "user_id": request.json['userID'],
+                "user_id": user_id,
                 "site_id": site_id,
                 "tag": request.json['tag'],
                 "views": 0,
@@ -251,7 +226,7 @@ def add_favorite():
             })
             return jsonify({"message":"Favorite added"})
     except Exception as e:
-        return jsonify({"message":"Could not add favorite", "error":str(e)}) # error code
+        return jsonify({"message":"Could not add favorite", "error":str(e)}), 500
 
 # Edit tag of favorite site
 @app.route('/edit-tag', methods=['POST'])
@@ -271,13 +246,15 @@ def edit_tag():
 # Remove favorite site
 @app.route('/remove-favorite', methods=['POST'])
 def remove_favorite():
-    # Authorization
+    user_id = check_authorization(request)
+    if user_id is None:
+        return jsonify({"message":"Unauthorized access"}), 403
     try:
-        favoriteSites.delete_one({ "_id": ObjectId(request.json['favorite_id'])}) # check if owned by user
+        favoriteSites.delete_one({ "_id": ObjectId(request.json['favorite_id']), "user_id": user_id })
         addedTo.delete_many({ "favorite_id": request.json['favorite_id']})
         return jsonify({"message":"Favorite removed"})
     except Exception as e:
-        return jsonify({"message":"Could not remove favorite", "error":str(e)}) # error code
+        return jsonify({"message":"Could not remove favorite", "error":str(e)}), 500
 
 ########################### ROUTES (LISTS) ###########################
 
@@ -296,20 +273,11 @@ def get_list():
 # Return all my lists
 @app.route('/my-lists', methods=['GET'])
 def get_my_lists():
-    auth_header = request.headers.get('Authorization')
-    if auth_header.startswith('Bearer '):
-        try:
-            encoded_jwt = auth_header[7:]
-            decoded_jwt = jwt.decode(encoded_jwt, os.getenv('JWT_KEY'), algorithms=['HS256'])
-            if 'username' not in decoded_jwt:
-                return jsonify({"error":"Unauthorized access"}), 403
-            user_id = decoded_jwt['username']
-            myLists = lists.find({"user_id": str(user_id) }).sort('name', 1)
-            return jsonify(myLists)
-        except Exception as e:
-            print("JWT decode error: ", e)
-            return jsonify({"error":"Unauthorized access"}), 403
-    return jsonify({"error":"Unauthorized access"}), 403
+    user_id = check_authorization(request)
+    if user_id is None:
+        return jsonify({"message":"Unauthorized access"}), 403
+    myLists = lists.find({"user_id": str(user_id) }).sort('name', 1)
+    return jsonify(myLists)
 
 # Return list items
 @app.route('/list-items', methods=['GET'])
@@ -335,60 +303,76 @@ def get_list_items():
 # Create new list
 @app.route('/create-list', methods=['POST'])
 def create_list():
-    # Authorization
+    user_id = check_authorization(request)
+    if user_id is None:
+        return jsonify({"message":"Unauthorized access"}), 403
     try:
-        duplicates_count = lists.count_documents({"user_id": request.json['userID'], "name": request.json['name']})
+        duplicates_count = lists.count_documents({"user_id": user_id, "name": request.json['name']})
         if duplicates_count > 0:
             return jsonify({"message":"You already have a list with this name"})
         else:
             lists.insert_one({
-                "user_id": request.json['userID'],
+                "user_id": user_id,
                 "name": request.json['name'],
                 "isPrivate": request.json['isPrivate'],
                 "dateAdded": str(datetime.datetime.now())[0:16]
             })
             return jsonify({"message":"List created"})
     except Exception as e:
-        return jsonify({"message":"Could not create list", "error":str(e)}) # error code
+        return jsonify({"message":"Could not create list", "error":str(e)}), 500
 
 # Remove list
 @app.route('/remove-list', methods=['POST'])
 def remove_list():
-    # Authorization
+    user_id = check_authorization(request)
+    if user_id is None:
+        return jsonify({"message":"Unauthorized access"}), 403
     try:
-        lists.delete_one({ "_id": ObjectId(request.json['id'])}) # check if owned by user
+        lists.delete_one({ "_id": ObjectId(request.json['id']), "user_id": user_id })
         addedTo.delete_many({ "list_id": request.json['id']})
         return jsonify({"message":"List removed"})
     except Exception as e:
-        return jsonify({"message":"Could not remove list", "error":str(e)}) # error code
+        return jsonify({"message":"Could not remove list", "error":str(e)}), 500
 
 # Add site to list
 @app.route('/add-to-list', methods=['POST'])
 def add_to_list():
-    # Authorization
+    user_id = check_authorization(request)
+    if user_id is None:
+        return jsonify({"message":"Unauthorized access"}), 403
     try:
-        duplicates_count = addedTo.count_documents({"favorite_id": request.json['favorite_id'], "list_id": request.json['list_id']})
+        list_id = request.json['list_id']
+        duplicates_count = addedTo.count_documents({"favorite_id": request.json['favorite_id'], "list_id": list_id})
         if duplicates_count > 0:
             return jsonify({"message":"Site already added to list"}) # error code
         else:
-            addedTo.insert_one({
-                "favorite_id": request.json['favorite_id'],
-                "list_id": request.json['list_id'], # check if list owned by user
-                "dateAdded": str(datetime.datetime.now())[0:16]
-            })
-            return jsonify({"message":"Site added to list"})
+            myList = lists.find_one({"_id":ObjectId(list_id), "user_id":user_id})
+            if myList is not None:
+                addedTo.insert_one({
+                    "favorite_id": request.json['favorite_id'],
+                    "list_id": list_id,
+                    "dateAdded": str(datetime.datetime.now())[0:16]
+                })
+                return jsonify({"message":"Site added to list"})
+            return jsonify({"message":"You can not modify somebody else's list"}) # error code
     except Exception as e:
-        return jsonify({"message":"Could not add site to list", "error":str(e)}) # error code
+        return jsonify({"message":"Could not add site to list", "error":str(e)}), 500
 
 # Remove site from list
 @app.route('/remove-from-list', methods=['POST'])
 def remove_from_list():
-    # Authorization
+    user_id = check_authorization(request)
+    if user_id is None:
+        return jsonify({"message":"Unauthorized access"}), 403
     try:
-        addedTo.delete_one({ "favorite_id": request.json['favorite_id'], "list_id": request.json['list_id']}) # check if list owned by user
-        return jsonify({"message":"Site removed from list"})
+        list_id = request.json['list_id']
+        myList = lists.find_one({"_id":ObjectId(list_id), "user_id":user_id})
+        if myList is not None:
+            addedTo.delete_one({ "favorite_id": request.json['favorite_id'], "list_id": list_id})
+            return jsonify({"message":"Site removed from list"})
+        return jsonify({"message":"You can not modify somebody else's list"}) # error code
     except Exception as e:
-        return jsonify({"message":"Could not remove site from list", "error":str(e)}) # error code
+        return jsonify({"message":"Could not remove site from list", "error":str(e)}), 500
 
 ########################### RUN APP ###########################
